change IdTruth from UShort_t to Int_t
Former-commit-id: 65d8cb0974123d50047d5f6d353bf4ae598aa995

+++ StRoot/StPxlSimMaker/StPxlFastSim.cxx
@@ -199,7 +199,7 @@ Int_t StPxlFastSim::addPxlHits(const StMcPxlHitCollection& mcPxlHitCol,
 
                UInt_t hw = sector * 10 + ladder; // needs to be updated later after clustering alogrithms are finalized
 
-               unsigned short idTruth = mcPix->parentTrack() ? mcPix->parentTrack()->key() : -999;
+               int idTruth = mcPix->parentTrack() ? mcPix->parentTrack()->key() : -999;
                unsigned short quality = mcPix->parentTrack() ? 100 : 0;
 
                StPxlDigiHit* tempHit = new StPxlDigiHit(localPixHitPos, iSec+1, mcPix->ladder(), mcPix->sensor(),
